+++ lib/src/lib/components/form-container/form-container.component.ts
@@ -8,6 +8,7 @@ import { Lab900FormField } from '../../models/lab900-form-field.type';
 import { EditType } from '../../models/editType';
 import { LAB900_FORM_MODULE_SETTINGS } from '../../models/Lab900FormModuleSettings';
 import { FormFieldDirective } from '../../directives/form-field.directive';
+import { uniqueId } from 'lodash';
 
 @Component({
   selector: 'lab900-form',
@@ -45,8 +46,8 @@ export class Lab900Form<T> {
     );
   });
   public readonly controls = computed(() => this._form().controls);
-  public readonly readonly = computed(() => this.schema()?.readonly);
-  public readonly formId = computed(() => this.schema()?.formId);
+  public readonly formIsReadOnly = computed(() => this.schema()?.readonly);
+  public readonly formId = computed(() => this.schema()?.formId ?? uniqueId());
 
   public get form(): UntypedFormGroup {
     return this._form();