+++ lib/src/lib/components/form-fields/search-field/search-field.component.ts
@@ -3,7 +3,7 @@ import { FormComponent } from '../../AbstractFormComponent';
 import { FormFieldSearch } from './field-search.model';
 import { take } from 'rxjs/operators';
 import { MatFormFieldModule } from '@angular/material/form-field';
-import { AsyncPipe, NgClass } from '@angular/common';
+import { AsyncPipe } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { TranslatePipe } from '@ngx-translate/core';
 import { MatInputModule } from '@angular/material/input';
@@ -22,7 +22,6 @@ import { MatProgressSpinner } from '@angular/material/progress-spinner';
     MatInputModule,
     ReactiveFormsModule,
     TranslatePipe,
-    NgClass,
     AutofocusDirective,
     SearchInputDirective,
     MatIconButton,