+++ lib/src/lib/components/form-fields/search-field/search-input.directive.ts
@@ -14,7 +14,6 @@ import { FormFieldSearchOptions } from './field-search.model';
       multi: true,
     },
   ],
-  standalone: true,
 })
 export class SearchInputDirective<T> implements ControlValueAccessor, OnChanges {
   private readonly searchQuery$ = new ReplaySubject<string>();