+++ lib/src/lib/services/form-builder.service.ts
@@ -8,7 +8,6 @@ import {
 } from '@angular/forms';
 import { EditType } from '../models/editType';
 import { inject, Injectable } from '@angular/core';
-import { FormFieldUtils } from '../utils/form-field.utils';
 import { Lab900FormField } from '../models/lab900-form-field.type';
 import { FormFieldAutocomplete } from '../components/form-fields/autocomplete-field/autocomplete-field.model';
 import { requireMatchValidator } from '../validators/require-match.validator';
@@ -18,16 +17,10 @@ import { FormFieldRepeater } from '../components/form-fields/repeater-field/repe
 export class Lab900FormBuilderService {
   private readonly fb = inject(UntypedFormBuilder);
 
-  public static addValidators(field: Lab900FormField, data: any): ValidatorFn[] {
+  public static addValidators(field: Lab900FormField): ValidatorFn[] {
     const validators: ValidatorFn[] = field?.validators ?? [];
     if (!field.options) {
       return [];
-    }
-    if (
-      !validators.includes(Validators.required) &&
-      FormFieldUtils.isRequired(FormFieldUtils.isReadOnly(field.options, data), field, data)
-    ) {
-      validators.push(Validators.required);
     }
     if (field.options?.minLength) {
       validators.push(Validators.minLength(field.options.minLength));
@@ -56,7 +49,7 @@ export class Lab900FormBuilderService {
       if (field.attribute) {
         if ((field.editType === EditType.Row || field.editType === EditType.Column) && field.nestedFields) {
           const nestedGroup = this.createFormGroup(field.nestedFields, undefined, data?.[field.attribute as keyof T]);
-          nestedGroup.setValidators(Lab900FormBuilderService.addValidators(field, data));
+          nestedGroup.setValidators(Lab900FormBuilderService.addValidators(field));
           formGroup.addControl(field.attribute, nestedGroup);
         } else {
           const fieldGroup = this.setFieldGroup(field.attribute, formGroup);
@@ -100,7 +93,7 @@ export class Lab900FormBuilderService {
 
     if (field.editType === EditType.Repeater) {
       const repeaterArray = this.createFormArray(formData, field);
-      repeaterArray.setValidators(Lab900FormBuilderService.addValidators(field, data));
+      repeaterArray.setValidators(Lab900FormBuilderService.addValidators(field));
       fieldGroup.addControl(attribute, repeaterArray);
     } else if (field.editType === EditType.DateRange) {
       const options = field?.options;
@@ -120,7 +113,7 @@ export class Lab900FormBuilderService {
             ? field.options.defaultValue(data)
             : field.options.defaultValue;
       }
-      const formControl = new UntypedFormControl(data, Lab900FormBuilderService.addValidators(field, data));
+      const formControl = new UntypedFormControl(data, Lab900FormBuilderService.addValidators(field));
       fieldGroup.addControl(attribute, formControl);
     }
   }