+++ src/app/modules/showcase-forms/examples/form-field-repeater-advanced-example/form-field-repeater-advanced-example.component.ts
@@ -31,15 +31,6 @@ export class FormFieldRepeaterAdvancedExampleComponent {
             editType: EditType.Row,
             options: {
               colspan: 12,
-              visibleFn: (item: any) => {
-                if (
-                  (item.group.parent.controls as any[]).indexOf(item.group) ===
-                  (item.group.parent.controls as any[]).length - 1
-                ) {
-                  return false;
-                }
-                return true;
-              },
             },
             nestedFields: [
               {